% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reference.R
\name{clean_hmdb_data}
\alias{clean_hmdb_data}
\title{Clean and Process HMDB Data}
\usage{
clean_hmdb_data(data)
}
\arguments{
\item{data}{Dataframe generated by the first steps of hmdbextract function}
}
\value{
Processed dataframe for the next steps.
}
\description{
Cleans and prepare the HMDB dataset by converting all columns to character, replacing NAs with empty strings,
removing unnecessary characters, and creating needed columns for the next steps.
}
\keyword{internal}
