% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reference.R
\name{hmdbextract}
\alias{hmdbextract}
\title{Extract metabolite information from HMDB XML File}
\usage{
hmdbextract(in_file)
}
\arguments{
\item{in_file}{A character string specifying the path to the HMDB XML file.}
}
\value{
A data frame containing extracted metabolite information with the following columns:
\itemize{
  \item \code{accession}: Primary accession ID.
  \item \code{secondary_accessions}: Secondary accession IDs, concatenated into a single string.
  \item \code{name}: Name of the metabolite.
  \item \code{description}: Description of the metabolite.
  \item \code{synonyms}: Synonyms for the metabolite, concatenated into a single string.
  \item \code{InChIKey}: InChIKey of the metabolite.
  \item \code{description_taxonomy}: Description of the taxonomy.
  \item \code{kingdom}: Taxonomic kingdom classification.
  \item \code{super_class}: Taxonomic super class.
  \item \code{class}: Taxonomic class.
  \item \code{sub_class}: Taxonomic subclass.
  \item \code{direct_parent}: Taxonomic direct parent classification.
  \item \code{chebi_id}: ChEBI ID of the metabolite.
  \item \code{kegg_id}: KEGG ID of the metabolite.
}
}
\description{
Parses an HMDB XML file to extract metabolite-related information, organizing it into a structured data frame that will be used as the HMDB reference for the next steps.
}
\details{
This function parses an HMDB XML file (XML can be downloaded from the "Downloads" section here :https://hmdb.ca). It iterates through each metabolite
node, extracts relevant fields, and returns the information as a data frame. The output will be used for the next steps as the reference.
}
\examples{
\dontrun{
  # Example usage
  extracted_data <- hmdbextract("path_to_hmdb.xml")
  head(extracted_data)
}

}
