% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_unique_identifier}
\alias{map_unique_identifier}
\title{Map Unique Identifiers}
\usage{
map_unique_identifier(df_cohort, col_cohort, df_reference, identifier)
}
\arguments{
\item{df_cohort}{A data frame containing cohort data with metabolite identifiers to be mapped.}

\item{col_cohort}{A character string specifying the column in `df_cohort` containing the identifier to be mapped.}

\item{df_reference}{A data frame containing reference data with metabolite identifiers and corresponding metadata.}

\item{identifier}{A character string specifying the type of identifier to map.
Supported values are `"INCHIKEY"`, `"NAME"`, `"HMDB"`, `"CHEBI"`, and `"KEGG"`.}
}
\value{
A data frame with updated mapping information, including matched identifiers
}
\description{
Maps a given metabolite identifier type in a cohort data frame to reference identifiers
    (e.g., INCHIKEY, NAME, HMDB, CHEBI, KEGG) by performing a series of processing, matching,
    and reporting steps.
}
\details{
- For `INCHIKEY`, both the full and skeleton INCHIKEY identifiers are mapped.
- For `NAME`, metabolite names are processed and matched to all known labels, inclusing synonyms.
- For `HMDB`, `CHEBI`, and `KEGG`, the corresponding identifiers are matched directly to reference identifiers.
}
\keyword{internal}
