% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{match_identifier}
\alias{match_identifier}
\title{Match Identifier}
\usage{
match_identifier(
  df_cohort,
  col_cohort,
  df_reference,
  col_reference,
  identifier
)
}
\arguments{
\item{df_cohort}{A data frame containing cohort data with metabolite identifiers to be mapped.}

\item{col_cohort}{A character string specifying the column in `df_cohort` containing the identifier to be mapped.}

\item{df_reference}{A data frame containing reference data with metabolite identifiers and corresponding metadata.}

\item{identifier}{A character string specifying the type of identifier to map.
Supported values are `"INCHIKEY"`, `"NAME"`, `"HMDB"`, `"CHEBI"`, and `"KEGG"`.}
}
\value{
A data frame that merges the cohort and reference data frames based on the specified identifier.Includes an additional column indicating whether a match was found.
}
\description{
Matches metabolite identifiers from a cohort data set with a reference database,
    annotating the matches and providing additional information.
}
\details{
- Counts and prints the number of available identifiers in the cohort data set before matching.
- Renames columns in the reference data frame by appending `"_in_reference_"` and the `identifier`
    to avoid naming conflicts.
- Merges the cohort and reference data frames based on the specified identifier columns.
- Adds a new logical column (`mapped_to_<col_reference>`) to indicate if a match was found.
}
\keyword{internal}
