% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboMapper_functions.R
\name{MetaboMap}
\alias{MetaboMap}
\title{MetaboMap}
\usage{
MetaboMap(
  df_study,
  study_cols,
  identifiers,
  df_reference,
  details = FALSE,
  original_colname
)
}
\arguments{
\item{df_study}{A data frame containing metabolite information from the study, where each row corresponds to a metabolite and each column contain one identifier.}

\item{study_cols}{A column name or a vector of column names in `df_study` corresponding to identifiers to be mapped.}

\item{identifiers}{An identifier or a vector of identifiers to be mapped; must be the same length and same order as `study_cols`. It can takes values c("HMDB","NAME","KEGG","CHEBI","INCHIKEY")}

\item{df_reference}{Dataframe generated by the function hmdbextract().}

\item{details}{A boolean- Default Value is FALSE. TRUE allows to have additional details for each mapping identifier.}

\item{original_colname}{A column name in `df_study` containing the original metabolites label in the study}
}
\value{
A data frame object containing the mapped identifiers with confidence levels.
}
\description{
Matches one or multiple metabolites identifiers of interest between the study dataframe and the generated reference dataframe from HMDB.
}
\examples{
study_data <- data.frame(
  metabolite_name = c("1-Methylhistidine", "2-Ketobutyric acid"),
  HMDB_id = c("HMDB0000001", "HMDB0000005")
)

reference_data <- data.frame(
  accession = c("HMDB0000001", "HMDB0000002", "HMDB0000005"),
  name = c("1-Methylhistidine", "1,3-Diaminopropane", "2-Ketobutyric acid"),
  all_HMDBs = c(
    "HMDB0000001|HMDB00001|HMDB0004935",
    "HMDB0000002|HMDB00002",
    "HMDB0000005|HMDB00005|HMDB0006544"
  ),
  all_names = c(
    "1-Methylhistidine|Pi-methylhistidine|",
    "1,3-Diaminopropane|1,3-Propanediamine|1,3-Propylenediamine",
    "2-Ketobutyric acid|2-Ketobutanoic acid|2-Oxobutyric acid"
  )
)

result <- MetaboMap(
  df_study = study_data,
  study_cols = c("metabolite_name", "HMDB_id"),
  identifiers = c("NAME", "HMDB"),
  df_reference = reference_data,
  details = FALSE,
  original_colname = "metabolite_name"
)
}
