% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_letters_between_parentheses}
\alias{remove_letters_between_parentheses}
\title{Remove Letters Between Parentheses}
\usage{
remove_letters_between_parentheses(text)
}
\arguments{
\item{text}{A character string containing text where letters within parentheses will be removed.}
}
\value{
A character string with letters inside parentheses removed. The parentheses themselves are also removed.
}
\description{
Removes any letters enclosed within parentheses from a given text string.
    It leaves the parentheses empty or removes them entirely if they contain only letters.
}
\details{
- The function uses a regular expression to match any sequence of letters inside parentheses and replace them with empty strings.
- It ensures that parentheses without letters are also cleaned, resulting in no empty parentheses remaining.
}
\keyword{internal}
