% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_stars_after_NA}
\alias{remove_stars_after_NA}
\title{Remove Stars After "NA"}
\usage{
remove_stars_after_NA(column)
}
\arguments{
\item{column}{A character vector representing a column of data. This can be a single
column from a dataframe or a standalone vector.}
}
\value{
A character vector with all occurrences of "NA" followed by one or more
asterisks (`*`) replaced with "NA".
}
\description{
This function removes all asterisks (`*`) that appear immediately after the string "NA"
in a given column of data. It uses a regular expression to identify and replace
such patterns.
}
\keyword{internal}
